requirements.txt
astroid==2.3.3
cachetools==3.1.1
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
future==0.18.2
google-api-python-client==1.7.11
google-auth==1.7.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
gspread==3.1.0
httplib2==0.14.0
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.17.4
oauth2client==4.1.3
oauthlib==3.1.0
pandas==0.25.3
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
pylint==2.4.4
python-dateutil==2.8.1
python-dotenv==0.10.3
python-telegram-bot==12.2.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.13.0
tornado==6.0.3
typed-ast==1.4.0
uritemplate==3.0.0
urllib3==1.25.7
wrapt==1.11.2
jinja2